Refactor: Split the Win32 platform layer into subsystems
